people_counter/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
mysql-connector-python>=8.0.0
numpy>=1.24.0
